RHCLOUD-23281 refactoring
